Support embedded created Response objects in ServiceWorker.
Response objects were previously created in Blink and provided to the
embedder; however the Cache API requires that they can cross the other
way. This CL adds the accessors required to do this; both to fill in a
Response object from a WebServiceWorkerResponse and to allow Blink to
read all the fields in a WebServiceWorkerResponse.

R=jkarlin@chromium.org,jsbell@chromium.org,asanka@chromium.org
BUG=399178

Review URL: https://codereview.chromium.org/519803002

git-svn-id: svn://svn.chromium.org/blink/trunk@181327 bbb929c8-8fbe-4397-9dbb-9b2b20218538