Merge FontPlatformDataHarfBuzz and FontPlatformData headers
Those two files were split apart and redundantly copied
for non-Mac and Mac platforms. This is an attempt an unifying
them, as a preparation for removing the CoreText shaper. 
We can merge the implementations in FontPlatformDataHarfBuzz.cpp 
and FontPlatformData.cpp as a next step.

BUG=334269

Review URL: https://codereview.chromium.org/516143003

git-svn-id: svn://svn.chromium.org/blink/trunk@181336 bbb929c8-8fbe-4397-9dbb-9b2b20218538