Mixed Content: Move subframe checks into ResourceFetcher.
Currently we're checking for mixed content when loading the main
resource of subframes in two places: DocumentLoader and ResourceFetcher.
The former properly checks against the frame's parent frame, while the
latter bypasses the checks entirely by virtue of the request's
'Resource::Type' being 'Main'.

This is both confusing and difficult to replicate when moving to the
brave new world of request contexts and frame types that Fetch now
defines.

This patch drops the DocumentLoader check, and moves the relevant
logic into ResourceFetcher::canRequest and
ResourceFetcher::checkInsecureContent.

BUG=400087

Review URL: https://codereview.chromium.org/544573002

git-svn-id: svn://svn.chromium.org/blink/trunk@181383 bbb929c8-8fbe-4397-9dbb-9b2b20218538