Move some scroll invalidations to the paint invalidation phase
Scrolling used to generate immediate invalidations that are now
postponed to the paint invalidation phase. In order to make this
works, scrolling updates needs to to happen before paint
invalidations.

The test changes are either neutral (same amount of invalidations,
except that we generate them one at a time) or
compositing/repaint/fixed-pos-inside-composited-intermediate-layer
is a progression.

Review URL: https://codereview.chromium.org/498773007

git-svn-id: svn://svn.chromium.org/blink/trunk@181475 bbb929c8-8fbe-4397-9dbb-9b2b20218538