[DirectWrite] Add support for font-weight suffixes
Add support for font family names with a font-weight suffix (such as
Light, Semibold, Black  etc). If the full name fails to match an 
installed font check for a font-weight suffix and if so strip it and set
the weight according to the prefix.

The current implementation matches IE when it comes to picking or
overriding the weight based on the suffix.

TEST=fast/text/font-weight-variant.html
BUG=335323, 410333, 410865, 408260
R=cpu@chromium.org, dglazkov@chromium.org

Review URL: https://codereview.chromium.org/542603004

git-svn-id: svn://svn.chromium.org/blink/trunk@181493 bbb929c8-8fbe-4397-9dbb-9b2b20218538