Temporary revert: Introduce include directive to make_private_script_source.py
Reason: As discussed with @haraken over IRC, we need to migrate to using chromium
resource, grd, system. In that we may need to think over how to use the include directive.

One of the harmony modules, [1], describes about using the 'import "module-name"' directive.
We could consider using that as well. As this needs some discussion, we decided to revert
this patch first and reconsider landing this once the formal decision is made about it.

R=haraken@chromium.org, yosin@chromium.org

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/import

Review URL: https://codereview.chromium.org/559493002

git-svn-id: svn://svn.chromium.org/blink/trunk@181640 bbb929c8-8fbe-4397-9dbb-9b2b20218538