Make slider thumb painting bail out when the node is null.
The other media control painting methods do handle a null node in
toParentMediaElement (as can happen with a pseudo-element, such as
::-webkit-scrollbar). The slider thumb painting should too.

There are currently unfounded assertions that this node is not null,
but author style can make it null, so a proper check is appropriate.

A layout test has been added for this case.

BUG=412154

Review URL: https://codereview.chromium.org/556453007

git-svn-id: svn://svn.chromium.org/blink/trunk@181656 bbb929c8-8fbe-4397-9dbb-9b2b20218538