Removed unregisterTerminationObserver from DatabaseContext::contextDestroyed
Method DatabaseContext::contextDestroyed always calls from ExecutionContext destructor. In the case when the context is WorkerGlobalScope - from WorkerGlobalScope destructor. This unregister call does nothing in this situation.
And it looks dangerous when we try to use the methods of the object in destructor. E.g. if we make the method ExecutionContext::isWorkerGlobalScope virtual the program will crash in DatabaseContext::contextDestroyed.

R=tkent@chromium.org,vsevik@chromium.org

Review URL: https://codereview.chromium.org/551373003

git-svn-id: svn://svn.chromium.org/blink/trunk@181711 bbb929c8-8fbe-4397-9dbb-9b2b20218538