Ouptut oldLocation and newLocation in trace of paint invalidation
I found that InvalidationLocationChange is a very frequent invalidation
reason for some repaint storms. The invalidation is because the
location (from the paint invalidation container) of the render object
changed since the last invalidation. It will be helpful to know how
the location changes when debugging repaint storms.

Review URL: https://codereview.chromium.org/561443002

git-svn-id: svn://svn.chromium.org/blink/trunk@181759 bbb929c8-8fbe-4397-9dbb-9b2b20218538