Correctly mark all mouse events derived from touch with FromTouch
Bug 401177 added more explicit tracking of the source of synthetic mouse
events.  This CL updates the remianing places that generate mouse event
from touch input to explicitly mark them as so.  To do this I've removed
an ASSERT I added that attempted to lock-in hit-test reduction
improvements.  We have a dedicated test for this (hit-test-counts.html)
and so the ASSERT wasn't adding much.  It's not worth the effort at this
point to refactor the touch drag and drop code (a little used feature)
to remove a couple of hit tests (but I've left the comments in place
saying that ideally we'd do so in the future).

BUG=408660

Review URL: https://codereview.chromium.org/558753002

git-svn-id: svn://svn.chromium.org/blink/trunk@181780 bbb929c8-8fbe-4397-9dbb-9b2b20218538