Shrink Document::parseQualifiedName a bit.
It duplicated a large code chunk by using a templated function full of 
generic code both for UChar and LChar. By extracting the generic code
to the caller the amount of duplicate code was greatly reduced. With clang
the function shrunk from 5 KB to 2.5 KB (which is still large but not crazy
large).

-----------------------------------------------------------------------------------------------------------------------
 -2196 - Source: /home/bratell/src/chromium/src/third_party/WebKit/Source/core/dom/Document.cpp - (gained 0, lost 2196)
-----------------------------------------------------------------------------------------------------------------------
  Shrunk symbols:
      -2196: blink::Document::parseQualifiedName(WTF::AtomicString const&, WTF::AtomicString&, WTF::AtomicString&, blink::ExceptionState&) type=t, (was 4808 bytes, now 2612 bytes)

BUG=
R=fs@opera.com

Review URL: https://codereview.chromium.org/555813003

git-svn-id: svn://svn.chromium.org/blink/trunk@182043 bbb929c8-8fbe-4397-9dbb-9b2b20218538