Fix behavior of label associated with control element
Continous clicking on label was not able to transfer the click
event to associated control element. This was because we were
not passing the events if there is selection on label.
    
Now, the behaviour of label element is as follows:
 - If there is double click, two clicks will be passed to control
   element. Control element will *not* be focused.
 - If there is selection of label element by dragging, no click
   event is passed. Also, no focus on control element.
 - If there is already a selection on label element and then label
   is clicked, then click event is passed to control element and
   control element is focused.
    
The above mentioned behaviour is same as Firefox.

BUG=411065

Review URL: https://codereview.chromium.org/556813002

git-svn-id: svn://svn.chromium.org/blink/trunk@182131 bbb929c8-8fbe-4397-9dbb-9b2b20218538